requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-import-export==3.2.0
blue==0.9.1
isort==5.12.0
djangorestframework_simplejwt==5.2.2
django-filter==23.2
python-dotenv==1.0.0
